requirements.txt
attrs==20.3.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
idna==2.5
importlib-metadata==4.0.1
iniconfig==1.1.1
kiwisolver==1.0.1
matplotlib==3.0.1
numpy==1.14.3
packaging==20.9
pandas==0.23.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.3.0
pytest==6.2.5
python-dateutil==2.7.3
pytz==2018.4
requests2==2.16.0
six==1.11.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.21.1
virtualenv==16.0.0
zipp==3.4.1
